bugprone-narrowing-conversions

+++ src/shuffle.cc
@@ -106,7 +106,7 @@ auto output_shuffled_fasta(std::vector<int> & deck,
   progress_init("Writing output", deck.size());
   auto counter = std::size_t{0};
   for(auto const sequence_id: deck) {
-    fasta_print_db_relabel(output_file, sequence_id, counter + 1);
+    fasta_print_db_relabel(output_file, sequence_id, static_cast<int>(counter + 1));
     progress_update(counter);
     ++counter;
   }
